added bongocat and via(sort of)support
