+++ keyboards/lotus58/keymaps/via/readme.md
@@ -1,0 +1,5 @@
+![Lotus58 Layout Image](https://i.imgur.com/76euDDU.png)
+
+# Default Lotus58 Layout
+
+This is the default layout for Lotus58. The only strange thing for this layout might be the placement of enter right down.
