+++ keyboards/lotus58/keymaps/via/rules.mk
@@ -1,0 +1,3 @@
+OLED_DRIVER_ENABLE = yes
+WPM_ENABLE = yes
+LTO_ENABLE = yes
